refactor: components Header and Footer
