+++ app/[lang]/layout.tsx
@@ -3,8 +3,8 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import Providers from "./providers";
 import { Locale, i18n } from "@/i18n.config";
-import Header from "@/components/header";
-import Footer from "@/components/footer";
+import Header from "@/components/Header1";
+import Footer from "@/components/Footer1";
 
 const inter = Inter({ subsets: ["latin"] });
